Spin: Move the Spin component to the core
[Issue] https://github.com/Samsung/TAU/issues/819
[Problem] Spin is used for the wearable profile only
[Solution] Move the Spin to the core component to use the base component for the time and date pickers
