+++ src/css/profile/mobile/common/spin.less
@@ -1,0 +1,39 @@
+.ui-spin {
+	position: relative;
+	display: inline-block;
+	height: 60 * @unit_base;
+	background-color: transparent;
+	overflow: hidden;
+	font-size: 22 * @unit_base;
+	padding: 0;
+	box-sizing: content-box;
+	width: 80 * @unit_base;
+	&-item {
+		background-color: transparent;
+		position: absolute;
+		left: 0;
+		top: 0;
+		width: 100%;
+		height: 30 * @unit_base;
+		display: flex;
+		flex-direction: column;
+		justify-content: center;
+		color: rgba(77, 207, 255, 1);
+		
+		&-selected {
+			font-size: 30 * @unit_base;
+			height: 30 * @unit_base;
+			color: rgba(255, 255, 255, 1);
+		}
+	}
+	&-enabling {
+		height: 100%;
+		.ui-spin-item {
+			transition: 300ms opacity linear;
+		}
+	}
+	&-placeholder {
+		opacity: 0;
+		pointer-events: none;
+	}
+}
