+++ src/css/profile/mobile/mobile.less
@@ -62,3 +62,4 @@
 @import "common/graph.less";
 @import "common/toggleslider.less";
 @import "common/coverflow.less";
+@import "common/spin.less";
