+++ src/js/mobile.js
@@ -117,6 +117,7 @@
 			"./profile/mobile/widget/IndexScrollbar",
 			"./profile/mobile/widget/Button",
 			"./profile/mobile/widget/FloatingActions",
+			"./profile/mobile/widget/Spin",
 			"./core/widget/core/ScrollHandler",
 			"./core/widget/core/Tabs",
 			"./profile/mobile/widget/GridView",
