+++ tests/js/core/widget/core/Spin/api/Spin.js
@@ -6,11 +6,11 @@
 
 		equal(typeof ns, "object", "Class tau exists");
 		equal(typeof ns.widget, "object", "Class tau.widget exists");
-		equal(typeof ns.widget.wearable, "object", "Class tau.widget.wearable exists");
-		equal(typeof ns.widget.wearable.Spin, "function", "Class tau.widget.wearable.Spin exists");
+		equal(typeof ns.widget.core, "object", "Class tau.widget.wearable exists");
+		equal(typeof ns.widget.core.Spin, "function", "Class tau.widget.wearable.Spin exists");
 
 		widget = ns.engine.instanceWidget(document.getElementById("spin"), "Spin");
-		Spin = ns.widget.wearable.Spin;
+		Spin = ns.widget.core.Spin;
 
 		equal(typeof widget.value, "function", "Method Spin.getStartValue exists");
 		equal(typeof widget.handleEvent, "function", "Method Spin.handleEvent exists");
