+++ tests/js/profile/mobile/widget/Spin/Spin.html
@@ -1,0 +1,22 @@
+<!DOCTYPE html>
+<html>
+	<head>
+		<meta name="viewport" content="width=device-width,user-scalable=no"/>
+		<title>Spin</title>
+		<meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
+		<link rel="stylesheet" href="../../../../../libs/qunit-1.11.0.css">
+		<link rel="stylesheet" href="../../../../../libs/dist/theme/default/tau.css">
+		<script src="../../../../../libs/qunit-1.11.0.js" type="text/javascript"></script>
+		<script src="../../../../../libs/support.js"></script>
+		<script src="../../../../../karma/tests/helpers.js"></script>
+		<script type="text/javascript" src="../../../../../libs/dist/js/tau.js"></script>
+	</head>
+
+	<body>
+		<!-- Start of page: #one -->
+		<div id="qunit"></div>
+		<div id="qunit-fixture">
+		</div>
+	</body>
+	<script src="Spin.js"></script>
+</html>
