+++ tests/js/profile/mobile/widget/Spin/test-data/sample.html
@@ -1,0 +1,1 @@
+<div id="spin" class="ui-spin sample-spin"></div>
