+++ tests/js/profile/wearable/widget/wearable/Spin/spin.html
@@ -3,35 +3,30 @@
 	<head>
 		<meta name="viewport" content="width=device-width,user-scalable=no"/>
 		<title>Spin</title>
-	<meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-	<link rel="stylesheet"
-		  href="../../../../../../libs/qunit-1.11.0.css">
-	<link rel="stylesheet"
-		  href="../../../../../../libs/dist/theme/default/tau.css">
-	<script src="../../../../../../libs/qunit-1.11.0.js"
-			type="text/javascript"></script>
-	<script src="../../../../../../libs/support.js"></script>
-	<script type="text/javascript"
-			src="../../../../../../libs/dist/js/tau.js"></script>
-	<script type="text/javascript">
+		<meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
+		<link rel="stylesheet" href="../../../../../../libs/qunit-1.11.0.css">
+		<link rel="stylesheet" href="../../../../../../libs/dist/theme/default/tau.css">
+		<script src="../../../../../../libs/qunit-1.11.0.js" type="text/javascript"></script>
+		<script src="../../../../../../libs/support.js"></script>
+		<script type="text/javascript" src="../../../../../../libs/dist/js/tau.js"></script>
+		<script type="text/javascript">
 			var tauConfig = {
 				autorun: false
 			};
-	</script>
+		</script>
 	</head>
-	
+
 	<body>
 		<!-- Start of page: #one -->
 		<div id="qunit"></div>
 		<div id="qunit-fixture">
 			<div class="ui-page" data-enable-page-scroll="false">
 				<div class="ui-content">
-					<div id="spin" 
-						class="ui-spin sample-spin">
-					</div>
+					<div id="spin" class="ui-spin sample-spin">
 				</div>
-			</div><!-- /page one -->
-		</div>
+			</div>
+		</div><!-- /page one -->
+	</div>
 	</body>
 	<script src="spin.js"></script>
 </html>
